+++ CotEditor/Sources/CETextFinder.m
@@ -50,9 +50,7 @@ static const NSUInteger kMaxHistorySize = 20;
 
 @property (nonatomic, nonnull) CEFindPanelController *findPanelController;
 @property (nonatomic, nonnull) NSNumberFormatter *integerFormatter;
-
-// readonly
-@property (readwrite, nonatomic, nonnull) NSColor *highlightColor;
+@property (nonatomic, nonnull) NSColor *highlightColor;
 
 #pragma mark Settings
 @property (readonly, nonatomic) BOOL usesRegularExpression;
@@ -345,11 +343,16 @@ static const NSUInteger kMaxHistorySize = 20;
     if (![self checkIsReadyToFind]) { return; }
     
     NSNumberFormatter *integerFormatter = [self integerFormatter];
-    NSDictionary *hightlightAttributes = @{NSBackgroundColorAttributeName: [self highlightColor]};
     NSTextView *textView = [self client];
     NSString *findString = [self sanitizedFindString];
     OGRegularExpression *regex = [self regex];
     NSArray<NSValue *> *scopeRanges = [self scopeRanges];
+    
+    NSUInteger numberOfGroups = [self usesRegularExpression] ? [regex numberOfGroups] + 1 : [regex numberOfGroups];
+    NSArray<NSColor *> *highlightColors = [self decomposeHighlightColorsInto:numberOfGroups];
+    if (![self usesRegularExpression]) {
+        highlightColors = [[highlightColors reverseObjectEnumerator] allObjects];
+    }
     
     NSRegularExpression *lineRegex = [NSRegularExpression regularExpressionWithPattern:@"\n" options:0 error:nil];
     NSString *string = [textView string];
@@ -363,6 +366,8 @@ static const NSUInteger kMaxHistorySize = 20;
             isCancelled = YES;
         }
     }];
+
+    [[textView layoutManager] removeTemporaryAttribute:NSBackgroundColorAttributeName forCharacterRange:NSMakeRange(0, [string length])];
     
     __weak typeof(self) weakSelf = self;
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
@@ -396,7 +401,23 @@ static const NSUInteger kMaxHistorySize = 20;
                 inlineRange.location -= lineRange.location;
                 NSString *lineString = [string substringWithRange:lineRange];
                 NSMutableAttributedString *lineAttrString = [[NSMutableAttributedString alloc] initWithString:lineString];
-                [lineAttrString addAttributes:hightlightAttributes range:inlineRange];
+                
+                for (NSUInteger i = 0; i < numberOfGroups; i++) {
+                    NSRange range = [match rangeOfSubstringAtIndex:i];
+                    
+                    if (range.length == 0) { continue; }
+                    
+                    NSColor *color = highlightColors[i];
+                    
+                    [lineAttrString addAttribute:NSBackgroundColorAttributeName value:color
+                                           range:NSMakeRange(range.location - lineRange.location, range.length)];
+                    
+                    // highlight string in text view
+                    dispatch_async(dispatch_get_main_queue(), ^{
+                        [[textView layoutManager] addTemporaryAttribute:NSBackgroundColorAttributeName value:color
+                                                      forCharacterRange:range];
+                    });
+                }
                 
                 [result addObject:@{CEFindResultRange: [NSValue valueWithRange:matchedRange],
                                     CEFindResultLineNumber: @(lineNumber),
@@ -974,6 +995,26 @@ static const NSUInteger kMaxHistorySize = 20;
              NSStringFromSelector(@selector(notBeginOfLineOption)),
              NSStringFromSelector(@selector(notEndOfLineOption)),
              ];
+}
+
+
+// ------------------------------------------------------
+/// create desired number of highlight colors from base highlight color
+- (nonnull NSArray<NSColor *> *)decomposeHighlightColorsInto:(NSUInteger)numberOfGroups
+// ------------------------------------------------------
+{
+    NSMutableArray<NSColor *> *highlightColors = [NSMutableArray arrayWithCapacity:numberOfGroups];
+    
+    CGFloat hue, saturation, brightness, alpha;
+    [[self highlightColor] getHue:&hue saturation:&saturation brightness:&brightness alpha:&alpha];
+    
+    for (CGFloat i = 0.0; i < numberOfGroups; i++) {
+        double dummy;
+        [highlightColors addObject:[NSColor colorWithCalibratedHue:modf(hue + i / numberOfGroups, &dummy)
+                                                        saturation:saturation brightness:brightness alpha:alpha]];
+    }
+    
+    return [highlightColors copy];
 }